LogServerManager supports storage-based log server
