+++ settings.gradle
@@ -4,7 +4,6 @@ include 'digdag-core'
 include 'digdag-spi'
 include 'digdag-client'
 include 'digdag-standards'
-include 'digdag-log-s3'
 include 'digdag-storage-s3'
 include 'digdag-server'
 include 'digdag-cli'
